refactor hooks to naming scheme
