+++ frontend/src/components/ResultsView/ItemMenu/PersonCard/ItemRow/ItemRow.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { didMount } from "../../../../../hooks/didMount";
+import { didMount } from "../../../../../hooks/didMount/didMount.hook";
 import {
   AppStore,
   Item,
