--- frontend/src/hooks/didClickAway/index.tsx
@@ -1,1 +1,0 @@
-export { didClickAway } from "./didClickAway";
