+++ frontend/src/hooks/didMount/didMount.hook.tsx
@@ -1,9 +1,11 @@
 import { useEffect, useRef } from "react";
 
-export const didMount = () => {
+const didMount = () => {
   const didMount = useRef(true);
   useEffect(() => {
     didMount.current = false;
   }, []);
   return didMount.current;
 };
+
+export { didMount };
